ft fix

+++ libft/ft_strdup.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: macassag <macassag@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hznty <hznty@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 15:08:43 by macassag          #+#    #+#             */
-/*   Updated: 2023/11/20 10:56:34 by macassag         ###   ########.fr       */
+/*   Updated: 2023/11/24 16:55:33 by hznty            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 char	*ft_strdup(const char *s)
 {
-	char			*cpy;
-	unsigned int	i;
+	char	*cpy;
+	size_t	i;
 
 	i = 0;
 	cpy = (char *) malloc((ft_strlen((char *) s) + 1) * (sizeof(char)));
 	if (!cpy)
 		return ((void *) 0);
-	while (s[i] && i < (unsigned int) ft_strlen((char *) s) + 1)
+	while (s[i] && i < ft_strlen((char *) s) + 1)
 	{
 		cpy[i] = s[i];
 		i++;
